Feat/backend properties api (#5)
* feat(domain): add agency and property image entities

* feat(dtos): add property DTOs

* feat(dtos): add lead create request and property list query DTOs

* feat(validation): add validation for leads and property queries

* feat(mapping): add AutoMappers

* feat(api): register validation and mapping services

* test(validation): cover lead and query validators

* ci(backend): add workflow for unit tests and run it

* ci(backend): rename workflows

* ci(backend): run unit tests on main branch only

* docs(backend): update README and Makefile to match current architecture

* chore(backend): improve architecture of backend

* refactor(domain): use property dto instead of mapping for properties

* feat(backend): implement crud + filters + paging for property

* feat(backend): Add properties controller CRUD + Swagger

* feat(mongo): extend property reprository

* feat(infra): fix backend Docker build and compose setup

* docs(makefile): update help and command descriptions

* chore(swagger): document properties endpoints

* tests(backend): add integration tests for properties api

* refactor(tests): centralize shared fixtures

* refactor(backend): centralize packages versions

* docs(backend): add README for tests; update main README to reflect changes

* chore(DevExp): standardize dev commands and scripts

* chore(ux): run make help without requiring extra permissions

* fix(docker): include package version files in build context

* fix(build): resolve FluentValidation version conflicts

* fix(ci): update Mongo config for integration tests

+++ backend/src/RealEstate.Domain/Entities/Property.cs
@@ -15,7 +15,7 @@ public sealed class Property
     public PropertyType Type { get; set; }
     public int Bedrooms { get; set; }
     public int Bathrooms { get; set; }
-    public double Area { get; set; }
+    public decimal Area { get; set; }
 
     public Guid BrokerId { get; set; }
     public PropertyStatus Status { get; set; }
